+++ changelog.md
@@ -6,3 +6,4 @@
 - Unit tests ported to JUnit 5
     - Group 1 unit tests ported
     - ListMatchHandlerTools enabled
+- Fixed dyed bundles and other custom recipes not populating in the Advanced Crafting Table via JEI or crafting book